chore: refactor search / route files in shared
