+++ dropbox-hacking-photo-manager-shared/src/ensureNever.ts
@@ -1,0 +1,3 @@
+export const ensureNever = (_: never) => {
+  throw new Error("ensureNever failed");
+};
