+++ dropbox-hacking-photo-manager-shared/src/index.ts
@@ -1,9 +1,10 @@
 export * from "./analyseAspectRatio.js";
 export * from "./conversation/index.js";
 export * from "./deltaFeed/index.js";
-export * from "./filter.js";
 export * from "./gpsLatLong.js";
+export * from "./routeState.js";
 export * from "./rxFeedTypes.js";
+export * from "./search/filterNode.js";
 export * from "./selectGPS.js";
 export * from "./types.js";
 export * from "./ws.js";
