--- dropbox-hacking-photo-manager-shared/src/search/execute.ts
@@ -1,12 +1,0 @@
-// const knownMediaExtensions = [
-//   ".jpg",
-//   ".png",
-//   ".mov",
-//   ".mp4",
-//   ".bmp",
-//   ".jpeg",
-//   ".cr3",
-// ] as const;
-
-// const knownMediaExtension = (name: string): boolean =>
-//   knownMediaExtensions.includes(extname(name).toLocaleLowerCase() as never);
