+++ dropbox-hacking-photo-manager-shared/src/search/filterNode.ts
